Update master_nano.ino
 At first the arduinos appeared to stop sending serial data soon after beginning communication. Increasing baud rate of the serial between the two microcontrollers and controlling how often commands are sent seemed to stop any hangings that were experienced before the changes were made. This suggests that serial buffer overflow may be an issue that warrants some consideration in the design of the protocol.

Fundamental changes to the protocol were made as well. The boards are now referred to as board 1 and board 2. They are differentiated from one another in software by reading the analog voltage input to A3 on the arduino mega. This pin is connected to a jumper on both boards. On board 1 it is connected to ground, on board 2 it is connected to 5V. This allows the same code to be uploaded to both microcontrollers to simplify operations and debugging. If else structures were placed in the code to set different timeout values for board 1 and board 2. These different timeout values were created to prevent transmit and receive modes on the boards from syncing in such a way that they both transmit while the other board switches between transmit and receiving modes. In addition an if else structure was placed in the loop of the arduino code that allows differentiation of which writeByCommands are called and when for board 1 and board 2. 

In addition the protocol was made much less recursive. writeByCommands are processed first in the arduino loop(){} section, then the end of stream packet is sent, and finally extractPackets() is called before the loop() is started again. Occasionally extractPackets() may cause an end of stream packet to be sent, in the case of a timeout, before resuming the extractPackets() function. The extractPackets() function will occasionally encounter packets that request a packet to be sent back to the other node, this will cause such requests to be added to the actOnBuff[]. When an end of stream packet is received the actOnBuff[] is flushed by flushActOnBuff(). Following the flushActOnBuff() call, the program returns to the start of the arduino loop(). The two cases described above are the only two possible ways an end of stream packet is sent. 

The code was also cleaned up a bit by removing watch dog timers, and old commented out code, that for now have no use because they don’t work. In the future we hope to add watch dog timers back to the code.